LayoutTests/http/tests/websocket/tests/hixie76/bufferedAmount-after-close.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Web Socket bufferedAmount after closed");

window.jsTestIsAsync = true;
if (window.layoutTestController)
    layoutTestController.overridePreference("WebKitHixie76WebSocketProtocolEnabled", 1);

var ws = new WebSocket("ws://localhost:8880/websocket/tests/hixie76/simple");

ws.onopen = function()
{
    debug("Connected.");
    ws.close();
};

ws.onclose = function()
{
    debug("Closed.");
    shouldBe("ws.readyState", "3");
    shouldBe("ws.bufferedAmount", "0");
    shouldBeFalse("ws.send('send to closed socket')");
    // If the connection is closed, bufferedAmount attribute's value will only
    // increase with each call to the send() method.
    // (the number does not reset to zero once the connection closes).
    shouldBe("ws.bufferedAmount", "23");
    finishJSTest();
};

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
